lectureGen/requirements.txt
flask==3.1.0
flask-cors==5.0.0
openai-whisper
anthropic
groq
scikit-learn
numpy
python-dotenv
werkzeug
